+++ src/Opaleye.hs
@@ -1,3 +1,5 @@
+-- {-# OPTIONS_HADDOCK ignore-exports #-}
+
 -- | An SQL-generating DSL targeting PostgreSQL.  Allows Postgres
 --   queries to be written within Haskell in a typesafe and composable
 --   fashion.
@@ -17,7 +19,6 @@ module Opaleye ( module Opaleye.Adaptors
                , module Opaleye.Aggregate
                , module Opaleye.Binary
                , module Opaleye.Column
-               , module Opaleye.Internal.Constant
                , module Opaleye.Distinct
                , module Opaleye.Field
                , module Opaleye.FunctionalJoin
@@ -28,9 +29,6 @@ module Opaleye ( module Opaleye.Adaptors
                , module Opaleye.MaybeFields
                , module Opaleye.Operators
                , module Opaleye.Order
-               , module Opaleye.Internal.PGTypesExternal
-               , module Opaleye.QueryArr
-               , module Opaleye.RunQuery
                , module Opaleye.RunSelect
                , module Opaleye.Sql
                , module Opaleye.Select
@@ -44,34 +42,19 @@ import Opaleye.Adaptors
 import Opaleye.Aggregate
 import Opaleye.Binary
 import Opaleye.Column
-import Opaleye.Internal.Constant
-  hiding (toFields)
+  hiding (null,
+          isNull)
 import Opaleye.Distinct
 import Opaleye.Field
-  hiding (null,
-          isNull,
-          matchNullable,
-          fromNullable,
-          toNullable,
-          maybeToNullable)
 import Opaleye.FunctionalJoin
 import Opaleye.Join
-  hiding (optional)
 import Opaleye.Label
 import Opaleye.Lateral
 import Opaleye.Manipulation
 import Opaleye.MaybeFields
-  hiding (optional)
 import Opaleye.Operators
 import Opaleye.Order
-import Opaleye.Internal.PGTypesExternal
-import Opaleye.QueryArr
-import Opaleye.RunQuery
 import Opaleye.RunSelect
-  hiding (foldForward,
-          closeCursor,
-          declareCursor,
-          declareCursorExplicit)
 import Opaleye.Select
 import Opaleye.Sql
 import Opaleye.SqlTypes