+++ src/Opaleye/Table.hs
@@ -15,7 +15,7 @@
  @
 
  The leftmost argument is the type of writes. When you insert or
- update into this column you must give it a @Field SqlInt4@ (which you
+ update into this field you must give it a @Field SqlInt4@ (which you
  can define with @sqlInt4 :: Int -> Field SqlInt4@).
 
  A required nullable @SqlInt4@ is defined with 'T.requiredTableField' and gives rise
@@ -25,7 +25,7 @@
  TableFields (FieldNullable SqlInt4) (FieldNullable SqlInt4)
  @
 
- When you insert or update into this column you must give it a
+ When you insert or update into this field you must give it a
  @FieldNullable SqlInt4@, which you can define either with @sqlInt4@ and
  @toNullable :: Field a -> FieldNullable a@, or with @null ::
  FieldNullable a@.
@@ -37,10 +37,10 @@
  TableFields (Maybe (Field SqlInt4)) (Field SqlInt4)
  @
 
- Optional columns are those that can be omitted on writes, such as
+ Optional fields are those that can be omitted on writes, such as
  those that have @DEFAULT@s or those that are @SERIAL@.
- When you insert or update into this column you must give it a @Maybe
- (Field SqlInt4)@. If you provide @Nothing@ then the column will be
+ When you insert or update into this field you must give it a @Maybe
+ (Field SqlInt4)@. If you provide @Nothing@ then the field will be
  omitted from the query and the default value will be used. Otherwise
  you have to provide a @Just@ containing a @Field SqlInt4@.
 
@@ -51,9 +51,9 @@
  TableFields (Maybe (FieldNullable SqlInt4)) (FieldNullable SqlInt4)
  @
 
- Optional columns are those that can be omitted on writes, such as
+ Optional fields are those that can be omitted on writes, such as
  those that have @DEFAULT@s or those that are @SERIAL@.
- When you insert or update into this column you must give it a @Maybe
+ When you insert or update into this field you must give it a @Maybe
  (FieldNullable SqlInt4)@. If you provide @Nothing@ then the default
  value will be used. Otherwise you have to provide a @Just@ containing
  a @FieldNullable SqlInt4@ (which can be null).
@@ -66,30 +66,21 @@ module Opaleye.Table (-- * Defining tables
                       T.Table,
                       T.tableField,
                       T.optionalTableField,
-                      T.readOnlyTableField,
                       T.requiredTableField,
+                      T.InferrableTableField,
                       -- * Selecting from tables
                       selectTable,
                       -- * Data types
-                      T.TableColumns,
                       TableFields,
                       -- * Explicit versions
                       selectTableExplicit,
-                      -- * Deprecated
-                      T.optional,
-                      T.readOnly,
-                      T.required,
-                      T.tableColumn,
-                      View,
-                      Writer,
-                      T.Table(T.Table, T.TableWithSchema),
-                      queryTable,
-                      queryTableExplicit) where
+                      -- * Deprecated versions
+                      T.readOnlyTableField,
+                     ) where
 
 import qualified Opaleye.Internal.QueryArr as Q
 import qualified Opaleye.Internal.Table as T
-import           Opaleye.Internal.Table (View, Table, Writer,
-                                         TableFields)
+import           Opaleye.Internal.Table (Table, TableFields)
 
 import qualified Opaleye.Internal.Tag as Tag
 import qualified Opaleye.Internal.Unpackspec as U
@@ -141,18 +132,7 @@ selectTableExplicit :: U.Unpackspec tablefields fields
                     -> Table a tablefields
                     -- ^
                     -> S.Select fields
-selectTableExplicit cm table' = Q.productQueryArr f where
-  f ((), t0) = (retwires, primQ, Tag.next t0) where
-    (retwires, primQ) = T.queryTable cm table' t0
-
--- * Deprecated versions
-
-{-# DEPRECATED queryTable "Use 'selectTable' instead.  Will be removed in version 0.8." #-}
-queryTable :: D.Default U.Unpackspec fields fields =>
-              Table a fields -> S.Select fields
-queryTable = selectTable
-
-{-# DEPRECATED queryTableExplicit "Use 'selectTableExplicit' instead.  Will be removed in version 0.8." #-}
-queryTableExplicit :: U.Unpackspec tablefields fields ->
-                     Table a tablefields -> S.Select fields
-queryTableExplicit = selectTableExplicit
+selectTableExplicit cm table' = Q.productQueryArr' $ \() -> do
+  t0 <- Tag.fresh
+  let (retwires, primQ) = T.queryTable cm table' t0
+  pure (retwires, primQ)