+++ src/Opaleye/TypeFamilies.hs
@@ -1,6 +1,4 @@
--- | Will be deprecated in 0.8.
-
-module Opaleye.TypeFamilies
+module Opaleye.TypeFamilies {-# DEPRECATED "No longer supported" #-}
   ( TF.TableRecordField
   , TF.RecordField
   , (TF.:<*>)
@@ -17,7 +15,6 @@ module Opaleye.TypeFamilies
   , TF.Opt
   , TF.Req
   , TF.Nulls
-  , TF.TableField
   ) where
 
 import Opaleye.Internal.TypeFamilies as TF